requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
SQLAlchemy==2.*
psycopg2-binary==2.9.*
alembic==1.*
python-dotenv==1.*
pydantic==2.*
pydantic-settings==2.*

